refactored tests for pytest, run with travis
