+++ python/xi_plugin.egg-info/SOURCES.txt
@@ -1,0 +1,12 @@
+setup.py
+xi_plugin/__init__.py
+xi_plugin/edit.py
+xi_plugin/linecache.py
+xi_plugin/plugin.py
+xi_plugin/rpc.py
+xi_plugin/style.py
+xi_plugin.egg-info/PKG-INFO
+xi_plugin.egg-info/SOURCES.txt
+xi_plugin.egg-info/dependency_links.txt
+xi_plugin.egg-info/not-zip-safe
+xi_plugin.egg-info/top_level.txt
