+++ python/xi_plugin.egg-info/not-zip-safe
@@ -1,0 +1,1 @@
+
